Rename dir to mockQueries
