+++ src/resolvers/tests/User.test.js
@@ -1,4 +1,4 @@
-import * as query from './queries/user';
+import * as query from './mockQueries/user';
 import { teams as mockTeams, users as mockUsers } from './mockData';
 import createTestServer from '../../testServer/createTestServer';
