+++ src/resolvers/tests/mockQueries/graphic.js
